ga4-proxy/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
google-analytics-data==0.18.0
gunicorn==21.2.0

